♻️ 데이터 fetching server에서 처리
