+++ components/activity-management/ActivityManagementCard.tsx
@@ -51,8 +51,8 @@ const ActivityManagementCard = ({ setActivities, prop }: Props) => {
           </Tooltip>
         </div>
         <div className="flex items-center justify-between ">
-          <span className=" text-base font-medium text-gray-80 md:text-2xl ">
-            ₩{price} /인
+          <span className=" text-sm font-medium text-gray-80 md:text-2xl ">
+            ₩{price} <span className="text-gray-50">/인</span>
           </span>
           <ActivityManagementKebab
             setActivities={setActivities}
